[venv] Add project directory to PYTHONPATH
Also delete '&&\' connection, and now checks existence of requirements
file.

+++ python_venv.sh
@@ -9,11 +9,20 @@ echo "-- Generate venv on ${PWD}/${venv_name}..."
 python3.6 -m venv ${venv_name} # by default python 3.6
 
 echo "-- Activate venv..."
-source ${venv_name}/bin/activate && \
+source ${venv_name}/bin/activate
 
-echo "-- Upgrade pip..." && \
-pip install --upgrade pip && \
+echo "-- Upgrade pip..."
+pip install --upgrade pip
 
-# TODO: check existance of requirements.txt
-echo "-- Add requirements..." && \
-pip install -r requirements.txt
+if [ -f requirements.txt ] ; then
+    echo "-- Add requirements..."
+    pip install -r requirements.txt
+else
+    echo "-- No requirements found..."
+fi
+
+if [ -z $PYTHONPATH ] ; then
+    export PYTHONPATH=$(pwd)
+else
+    export PYTHONPATH="$PYTHONPATH:$(pwd)"
+fi
